@@ -3,8 +3,6 @@ from apps.demo.models import *
 
 
 class MovieIn(ModelSchema):
-    
-
     class Meta:
         model = Movie
         fields = ['title', 'description', 'year', 'rating', 'genre', 'director', 'actors', ]
@@ -13,4 +11,4 @@ class MovieIn(ModelSchema):
 class MovieOut(ModelSchema):
     class Meta:
         model = Movie
-        fields = ['id', 'title', 'description', 'year', 'rating', 'genre', 'director', 'actors', ]
+        fields = ['id', 'is_deleted', 'created_time', 'updated_time', 'title', 'description', 'year', 'rating', 'genre', 'director', 'actors', ]
