@@ -1,4 +1,4 @@
-# Generated by Django 5.0.3 on 2024-06-05 15:23
+# Generated by Django 5.2 on 2025-05-14 10:38
 
 import django.db.models.deletion
 from django.db import migrations, models
@@ -13,41 +13,48 @@ class Migration(migrations.Migration):
 
     operations = [
         migrations.CreateModel(
-            name='MusicAlbum',
-            fields=[
-                ('id', models.BigAutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
-                ('name', models.CharField(max_length=100)),
-                ('year', models.IntegerField()),
-            ],
-            options={
-                'verbose_name': 'MusicAlbum',
-                'verbose_name_plural': 'MusicAlbums',
-                'db_table': 'demo_music_album',
-                'ordering': ['name'],
-            },
-        ),
-        migrations.CreateModel(
             name='Actor',
             fields=[
                 ('id', models.BigAutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
+                ('is_deleted', models.BooleanField(default=False, verbose_name='软删除标志')),
+                ('created_time', models.DateTimeField(auto_now_add=True, verbose_name='创建时间')),
+                ('updated_time', models.DateTimeField(auto_now=True, verbose_name='更新时间')),
                 ('name', models.CharField(max_length=100)),
                 ('birth_date', models.DateField()),
                 ('country', models.CharField(max_length=100)),
                 ('city', models.CharField(max_length=100)),
             ],
             options={
-                'verbose_name': 'Actor',
-                'verbose_name_plural': 'Actors',
+                'verbose_name': '演员',
+                'verbose_name_plural': '演员',
                 'db_table': 'demo_actor',
                 'ordering': ['name'],
-                'unique_together': {('name', 'name')},
-                'index_together': {('name', 'name')},
+            },
+        ),
+        migrations.CreateModel(
+            name='MusicAlbum',
+            fields=[
+                ('id', models.BigAutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
+                ('is_deleted', models.BooleanField(default=False, verbose_name='软删除标志')),
+                ('created_time', models.DateTimeField(auto_now_add=True, verbose_name='创建时间')),
+                ('updated_time', models.DateTimeField(auto_now=True, verbose_name='更新时间')),
+                ('name', models.CharField(max_length=100)),
+                ('year', models.IntegerField()),
+            ],
+            options={
+                'verbose_name': '音乐专辑',
+                'verbose_name_plural': '音乐专辑',
+                'db_table': 'demo_music_album',
+                'ordering': ['name'],
             },
         ),
         migrations.CreateModel(
             name='Movie',
             fields=[
                 ('id', models.BigAutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
+                ('is_deleted', models.BooleanField(default=False, verbose_name='软删除标志')),
+                ('created_time', models.DateTimeField(auto_now_add=True, verbose_name='创建时间')),
+                ('updated_time', models.DateTimeField(auto_now=True, verbose_name='更新时间')),
                 ('title', models.CharField(max_length=100)),
                 ('description', models.TextField()),
                 ('year', models.IntegerField()),
@@ -57,18 +64,21 @@ class Migration(migrations.Migration):
                 ('actors', models.CharField(max_length=100)),
             ],
             options={
-                'verbose_name': 'Movie',
-                'verbose_name_plural': 'Movies',
+                'verbose_name': '电影',
+                'verbose_name_plural': '电影',
                 'db_table': 'demo_movie',
                 'ordering': ['title'],
+                'indexes': [models.Index(fields=['title', 'year'], name='demo_movie_title_50f2dc_idx')],
                 'unique_together': {('title', 'year')},
-                'index_together': {('title', 'year')},
             },
         ),
         migrations.CreateModel(
             name='Music',
             fields=[
                 ('id', models.BigAutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
+                ('is_deleted', models.BooleanField(default=False, verbose_name='软删除标志')),
+                ('created_time', models.DateTimeField(auto_now_add=True, verbose_name='创建时间')),
+                ('updated_time', models.DateTimeField(auto_now=True, verbose_name='更新时间')),
                 ('name', models.CharField(max_length=100)),
                 ('singer', models.CharField(max_length=100)),
                 ('genre', models.CharField(max_length=100)),
@@ -76,8 +86,8 @@ class Migration(migrations.Migration):
                 ('album', models.ForeignKey(on_delete=django.db.models.deletion.CASCADE, to='demo.musicalbum')),
             ],
             options={
-                'verbose_name': 'Music',
-                'verbose_name_plural': 'Music',
+                'verbose_name': '音乐',
+                'verbose_name_plural': '音乐',
                 'db_table': 'demo_music',
                 'ordering': ['name'],
             },
