@@ -1,8 +1,10 @@
 from django.db import models
+from django_starter.db.models import ModelExt
+from simple_history.models import HistoricalRecords
 
 
 # Create your models here.
-class Movie(models.Model):
+class Movie(ModelExt):
     title = models.CharField(max_length=100)
     description = models.TextField()
     year = models.IntegerField()
@@ -10,6 +12,7 @@ class Movie(models.Model):
     genre = models.CharField(max_length=100)
     director = models.CharField(max_length=100)
     actors = models.CharField(max_length=100)
+    history = HistoricalRecords()
 
     def __str__(self):
         return self.title
@@ -17,17 +20,20 @@ class Movie(models.Model):
     class Meta:
         db_table = 'demo_movie'
         ordering = ['title']
-        verbose_name = 'Movie'
-        verbose_name_plural = 'Movies'
+        verbose_name = '电影'
+        verbose_name_plural = verbose_name
         unique_together = (('title', 'year'),)
-        index_together = (('title', 'year'),)
+        indexes = [
+            models.Index(fields=["title", "year"]),
+        ]
 
 
-class Actor(models.Model):
+class Actor(ModelExt):
     name = models.CharField(max_length=100)
     birth_date = models.DateField()
     country = models.CharField(max_length=100)
     city = models.CharField(max_length=100)
+    history = HistoricalRecords()
 
     def __str__(self):
         return self.name
@@ -35,15 +41,14 @@ class Actor(models.Model):
     class Meta:
         db_table = 'demo_actor'
         ordering = ['name']
-        verbose_name = 'Actor'
-        verbose_name_plural = 'Actors'
-        unique_together = (('name', 'name'),)
-        index_together = (('name', 'name'),)
+        verbose_name = '演员'
+        verbose_name_plural = verbose_name
 
 
-class MusicAlbum(models.Model):
+class MusicAlbum(ModelExt):
     name = models.CharField(max_length=100)
     year = models.IntegerField()
+    history = HistoricalRecords()
 
     def __str__(self):
         return self.name
@@ -51,16 +56,17 @@ class MusicAlbum(models.Model):
     class Meta:
         db_table = 'demo_music_album'
         ordering = ['name']
-        verbose_name = 'MusicAlbum'
-        verbose_name_plural = 'MusicAlbums'
+        verbose_name = '音乐专辑'
+        verbose_name_plural = verbose_name
 
 
-class Music(models.Model):
+class Music(ModelExt):
     name = models.CharField(max_length=100)
     singer = models.CharField(max_length=100)
     genre = models.CharField(max_length=100)
     rating = models.IntegerField()
     album = models.ForeignKey('MusicAlbum', on_delete=models.CASCADE)
+    history = HistoricalRecords()
 
     def __str__(self):
         return self.name
@@ -68,5 +74,5 @@ class Music(models.Model):
     class Meta:
         db_table = 'demo_music'
         ordering = ['name']
-        verbose_name = 'Music'
-        verbose_name_plural = 'Music'
+        verbose_name = '音乐'
+        verbose_name_plural = verbose_name
