@@ -3,10 +3,11 @@ from typing import Any, Mapping
 import orjson
 from django.conf import settings
 from django.http import HttpRequest
-from ninja import NinjaAPI
+from ninja import NinjaAPI, Swagger
 from ninja.renderers import JSONRenderer, BaseRenderer
 from django_starter.apis import router
 from apps.account.apis import router as account_router
+from apps.billing.apis import router as billing_router
 from apps.demo.apis import router as demo_router
 
 
@@ -32,8 +33,10 @@ api = NinjaAPI(
     description=settings.DJANGO_STARTER["project_info"]["description"],
     renderer=ORJSONRenderer(),
     urls_namespace='api',
+    docs=Swagger(settings={"persistAuthorization": True})
 )
 
 api.add_router('django-starter', router)
 api.add_router('account', account_router)
+api.add_router('billing', billing_router)
 api.add_router('demo', demo_router)
