@@ -1,7 +1,8 @@
 from django.urls import path, include
-from django_starter.contrib.guide import views
 
 urlpatterns = [
-    path('', views.index),
+    path('', include('django_starter.contrib.guide.urls')),
+    path('docs/', include('django_starter.contrib.docs.urls')),
+    path('notifications/', include('django_starter.contrib.notifications.urls')),
     path('admin/', include('django_starter.contrib.admin.urls')),
 ]
